Refactoring telegram bot code: onboarding handlers moved to onboarding.py, added mock command handlers
