+++ llm/llm_config.yaml
@@ -1,4 +1,4 @@
-llm_provider: "none"  # Options: google, openai, none (for debugging)
+llm_provider: "none"  # Options: google, openai, yandex or none (for debugging)
 google:
   model: "gemini-2.0-flash"
   api_key_env: "GOOGLE_API_KEY"
@@ -6,3 +6,7 @@ google:
 openai:
   model: "gpt-4"
   api_key_env: "OPENAI_API_KEY"
+
+yandex:
+  model: ""
+  iam_token: "IAM_TOKEN"
